Start tracking `less` updates
This workflow tracks the `Last-Modified` header's value as well as the
full contents of any given website in the repository, and if an update
is found, applies a sed script to extract the relevant information. If
new information was found, the workflow open a new issue.

The logic to avoid opening duplicate issues was copy/edited from the
`rss-to-issues` Action.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>