Align issue title format with regular component updates
Over in git-for-windows/git, we already have a workflow that opens
tickets when newer versions of some of Git for Windows' components are
available.

These tickets are picked up by the automation provided by the
GitForWindowsHelper GitHub App.

Let's align with the format so that we can eventually teach
GitForWindowsHelper to pick these tickets up, too.

Or maybe we should instead change the `track-website-changes` workflow
to open tickets over in git-for-windows/git? That's potentially a better
idea.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>